Extract ThankYouContext from CheckoutContext
