+++ src/Sylius/Behat/Context/Ui/Shop/Checkout/ThankYouContext.php
@@ -1,0 +1,67 @@
+<?php
+
+namespace Sylius\Behat\Context\Ui\Shop\Checkout;
+
+use Behat\Behat\Context\Context;
+use Sylius\Behat\Page\Shop\Order\ThankYouPageInterface;
+use Sylius\Component\Core\Model\PaymentMethodInterface;
+use Webmozart\Assert\Assert;
+
+/**
+ * @author Kamil Kokot <kamil.kokot@lakion.com>
+ */
+final class ThankYouContext implements Context
+{
+    /**
+     * @var ThankYouPageInterface
+     */
+    private $thankYouPage;
+
+    /**
+     * @param ThankYouPageInterface $thankYouPage
+     */
+    public function __construct(ThankYouPageInterface $thankYouPage)
+    {
+        $this->thankYouPage = $thankYouPage;
+    }
+
+    /**
+     * @Then I should see the thank you page
+     */
+    public function iShouldSeeTheThankYouPage()
+    {
+        Assert::true($this->thankYouPage->hasThankYouMessage());
+    }
+
+    /**
+     * @Then I should not see the thank you page
+     */
+    public function iShouldNotSeeTheThankYouPage()
+    {
+        Assert::false($this->thankYouPage->isOpen());
+    }
+
+    /**
+     * @Given I should be informed with :paymentMethod payment method instructions
+     */
+    public function iShouldBeInformedWithPaymentMethodInstructions(PaymentMethodInterface $paymentMethod)
+    {
+        Assert::same($this->thankYouPage->getInstructions(), $paymentMethod->getInstructions());
+    }
+
+    /**
+     * @Then I should not see any instructions about payment method
+     */
+    public function iShouldNotSeeAnyInstructionsAboutPaymentMethod()
+    {
+        Assert::false($this->thankYouPage->hasInstructions());
+    }
+
+    /**
+     * @Then I should not be able to change payment method
+     */
+    public function iShouldNotBeAbleToChangeMyPaymentMethod()
+    {
+        Assert::false($this->thankYouPage->hasChangePaymentMethodButton());
+    }
+}
