+++ src/Sylius/Behat/Resources/config/services/contexts/ui.xml
@@ -274,13 +274,17 @@
             <argument type="service" id="sylius.behat.page.shop.home" />
             <argument type="service" id="sylius.behat.page.shop.checkout.address" />
             <argument type="service" id="sylius.behat.page.shop.checkout.select_payment" />
-            <argument type="service" id="sylius.behat.page.shop.order.thank_you" />
             <argument type="service" id="sylius.behat.page.shop.checkout.select_shipping" />
             <argument type="service" id="sylius.behat.page.shop.checkout.complete" />
             <argument type="service" id="sylius.behat.page.shop.order.show" />
             <argument type="service" id="__symfony__.sylius.factory.address" />
             <argument type="service" id="sylius.behat.current_page_resolver" />
             <argument type="service" id="__symfony__.sylius.address_comparator" />
+            <tag name="fob.context_service" />
+        </service>
+
+        <service id="sylius.behat.context.ui.checkout.thank_you" class="Sylius\Behat\Context\Ui\Shop\Checkout\ThankYouContext">
+            <argument type="service" id="sylius.behat.page.shop.order.thank_you" />
             <tag name="fob.context_service" />
         </service>
