+++ src/Sylius/Behat/Resources/config/suites/ui/inventory/checkout_inventory.yml
@@ -27,6 +27,7 @@ default:
 
                 - sylius.behat.context.ui.shop.cart
                 - sylius.behat.context.ui.checkout
+                - sylius.behat.context.ui.checkout.thank_you
                 - sylius.behat.context.ui.admin.managing_product_variants
             filters:
                 tags: "@checkout_inventory && @ui"
